code/worker/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.4.26
click==8.2.1
dotenv==0.9.9
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
numpy==2.2.6
ollama==0.4.8
opencv-python==4.11.0.86
packaging==25.0
pdf2image==1.17.0
pillow==11.2.1
pydantic==2.11.5
pydantic_core==2.33.2
pytesseract==0.3.13
python-dotenv==1.1.0
redis==6.1.0
rq==2.3.3
sniffio==1.3.1
typing-inspection==0.4.1
typing_extensions==4.13.2
